+++ apps/server/src/api/products/analytics/analytics.controller.ts
@@ -1,17 +1,17 @@
-import { Controller, Post, Body, Delete, Query, BadRequestException, NotFoundException, Get } from "@nestjs/common"
-import { InsightsService } from "./insights.service"
+import { Controller, Post, Body, BadRequestException, Get } from "@nestjs/common"
 import { CreateAnalyticsDto } from "./dto/create-analytics.dto"
 import { CredentialAuthorizer, CredentialAuthorizerResponse } from "src/authorization/credential-authorizer.decorator"
 import { TokenAuthorizer, TokenAuthorizerResponse } from "src/authorization/token-authorizer.decorator"
+import { AnalyticsService } from "./analytics.service"
 
-@Controller("products/insights")
-export class InsightsController {
-  constructor(private readonly insightsService: InsightsService) { }
+@Controller("products/analytics")
+export class AnalyticsController {
+  constructor(private readonly analyticsService: AnalyticsService) { }
 
-  @Get("getanalytics")
+  @Get("get")
   async getAnalytics(@TokenAuthorizer() uft: TokenAuthorizerResponse) {
     try {
-      return await this.insightsService.getAnalytics(uft.workspaceId)
+      return await this.analyticsService.getAnalytics(uft.workspaceId)
     }
 
     catch (error) {
@@ -19,10 +19,10 @@ export class InsightsController {
     }
   }
 
-  @Post("createanalytics")
+  @Post("create")
   async createAnalytics(@CredentialAuthorizer() ufc: CredentialAuthorizerResponse, @Body() createAnalyticsDto: CreateAnalyticsDto) {
     try {
-      return await this.insightsService.createAnalytics(ufc.workspaceId, createAnalyticsDto)
+      return await this.analyticsService.createAnalytics(ufc.workspaceId, createAnalyticsDto)
     }
 
     catch (error) {