+++ apps/server/src/api/products/analytics/models/analytics.model.ts
@@ -1,0 +1,4 @@
+import { analyticsMongoDbConn } from "src/lib/db-connect"
+import { AnalyticsSchema } from "../schemas/analytics.schema"
+
+export const AnalyticsModel = analyticsMongoDbConn.model("analytics", AnalyticsSchema)