+++ apps/server/src/api/user/utils/send-email.ts
@@ -3,8 +3,8 @@ import { google } from "googleapis"
 import { envConfig } from "../../../env.config"
 import SMTPTransport from "nodemailer/lib/smtp-transport"
 
-const { gcloudClientId, gcloudClientSecret, redirectUri, refreshToken, mailerEmail } = envConfig
-const oAuth2Client = new google.auth.OAuth2(gcloudClientId, gcloudClientSecret, redirectUri)
+const { gcloudClientId, gcloudClientSecret, redirectURI, refreshToken, mailerEmail } = envConfig
+const oAuth2Client = new google.auth.OAuth2(gcloudClientId, gcloudClientSecret, redirectURI)
 oAuth2Client.setCredentials({ refresh_token: refreshToken })
 
 export async function sendEmail(email: string, passKey: string): Promise<void> {