+++ apps/server/src/lib/db-connect.ts
@@ -1,11 +1,11 @@
 import mongoose, { Connection } from "mongoose"
 import { envConfig } from "../env.config"
 
-export const platformMongoDbConn: Connection = mongoose.createConnection(envConfig.platformDbUri)
+export const platformMongoDbConn: Connection = mongoose.createConnection(envConfig.platformDbURI)
+export const analyticsMongoDbConn: Connection = mongoose.createConnection(envConfig.analyticsDbURI)
 export const copilotMongoDbConn: Connection = mongoose.createConnection(envConfig.copilotDbURI)
 export const datalakeMongoDbConn: Connection = mongoose.createConnection(envConfig.datalakeDbURI)
 export const walletMongoDbConn: Connection = mongoose.createConnection(envConfig.walletDbURI)
-export const insightsMongoDbConn: Connection = mongoose.createConnection(envConfig.insightsDbURI)
 export const nftstudioMongoDbConn: Connection = mongoose.createConnection(envConfig.nftstudioDbURI)
 export const swapMongoDbConn: Connection = mongoose.createConnection(envConfig.swapDbURI)
 export const fabricMongoDbConn = mongoose.createConnection(envConfig.fabricDbURI)
@@ -14,6 +14,11 @@ export const ledgerscanMongoDbConn: Connection = mongoose.createConnection(envCo
 async function platformDbConnect(): Promise<void> {
   platformMongoDbConn.on("connected", () => console.log("Platform DB Connected"))
   platformMongoDbConn.on("error", () => console.log("Platform DB Not Connected"))
+}
+
+async function analyticsDbConnect(): Promise<void> {
+  analyticsMongoDbConn.on("connected", () => console.log("Analytics DB Connected"))
+  analyticsMongoDbConn.on("error", () => console.log("Analytics DB Not Connected"))
 }
 
 async function copilotDbConnect(): Promise<void> {
@@ -29,11 +34,6 @@ async function datalakeDbConnect(): Promise<void> {
 async function walletDbConnect(): Promise<void> {
   walletMongoDbConn.on("connected", () => console.log("Wallet DB Connected"))
   walletMongoDbConn.on("error", () => console.log("Wallet DB Not Connected"))
-}
-
-async function insightsDbConnect(): Promise<void> {
-  insightsMongoDbConn.on("connected", () => console.log("Insights DB Connected"))
-  insightsMongoDbConn.on("error", () => console.log("Insights DB Not Connected"))
 }
 
 async function nftstudioDbConnect(): Promise<void> {
@@ -58,10 +58,10 @@ async function ledgerscanDbConnect(): Promise<void> {
 
 export async function dbConnect(): Promise<void> {
   platformDbConnect()
+  analyticsDbConnect()
   copilotDbConnect()
   datalakeDbConnect()
   walletDbConnect()
-  insightsDbConnect()
   nftstudioDbConnect()
   swapDbConnect()
   fabricDbConnect()