+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### v10.0.0
 * Added type safety for dispatcher in context
+* Deprecation of Insights 
+* Introduction of Analytics with new CQRS architecture
 
 ### v9.3.1
 * Minor bug fixes & performance improvements